Update CMakeLists.txt
Currently the find_package for mbedtls will always find one pre-installed on system (i.e. on mac if you have one installed via homebrew it will pick up that one) but if one has been installed in a local build environment you cannot specify it.  This is to test it will find it if it's there.